migrate amount field from primitive types to BigDecimal and adjust related classes and validations

+++ src/main/java/com/task/decisionengine/domain/LoanOffer.java
@@ -2,10 +2,12 @@ package com.task.decisionengine.domain;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record LoanOffer(
         DecisionOutcome outcome,
-        int amount,
+        BigDecimal amount,
         int period
 ) {
 }
